requirements-dev.txt
pytest>=7.0.0
PyYAML>=6.0
PySide6>=6.9.3